feat(CreateProductUseCase): extracting post handler to use case class
